API/requirements.txt
requests
aiohttp
numpy
python-dotenv
runpod
